[micromanage] Add schemas
This commit adds an initial draft of schemas. If an entity is created
without a schema, it fails. Additionally, if a property does not match
the schema, it fails.

I like the strictness, but I also know that it would be a bummer for
an API to change a string or number field, or decide to return null. I
still need to figure out the best approach for that.

One idea is to allow it, with an option to send warnings about schema
mismatches in a backchannel. You'd get an alert about the issue, but a
user wouldn't see an error.

This is probably okay for numbers, strings, and booleans. What about
dates?